move : Buy Mapper 분리
