+++ src/main/java/bc1/gream/domain/buy/mapper/BuyMapper.java
@@ -1,8 +1,8 @@
-package bc1.gream.domain.order.mapper;
+package bc1.gream.domain.buy.mapper;
 
-import bc1.gream.domain.order.dto.response.BuyBidResponseDto;
-import bc1.gream.domain.order.dto.response.BuyNowResponseDto;
-import bc1.gream.domain.order.entity.Buy;
+import bc1.gream.domain.buy.dto.response.BuyBidResponseDto;
+import bc1.gream.domain.buy.dto.response.BuyNowResponseDto;
+import bc1.gream.domain.buy.entity.Buy;
 import bc1.gream.domain.order.entity.Order;
 import bc1.gream.domain.product.dto.response.TradeResponseDto;
 import org.mapstruct.Mapper;
